+++ src/mathFun.c
@@ -1,3 +1,8 @@
+/**
+ * \file mathFun.c
+ * this is where the mathematical functions are defined
+ */
+
 #include <math.h>
 #include <stdlib.h>
 #include "mathFun.h"
@@ -8,14 +13,14 @@ float sigmoid(float input) {
     return 1/partial;
 }
 
-float ReLU(short input) {
+float ReLU(float input) {
     return fmax(0, input);
 }
 
 float randWeight(signed int min, signed int max) {
     int magicNum = 128;
     int buffMin = min * magicNum; int buffMax = max * magicNum; 
-    signed int randNum = rand()%buffMax;
+    float randNum = rand()%buffMax;
     randNum += buffMin;
     randNum /= 128;
     return randNum*((min+max)/2);
